+++ lte/phy/include/liblte/phy/phch/regs.h
@@ -45,23 +45,23 @@
 typedef _Complex float cf_t;
 
 typedef struct LIBLTE_API {
-  uint16_t k[4];
-  uint16_t k0;
-  uint8_t l;
+  uint32_t k[4];
+  uint32_t k0;
+  uint32_t l;
   bool assigned;
 }regs_reg_t;
 
 typedef struct LIBLTE_API {
-  uint16_t nof_regs;
+  uint32_t nof_regs;
   regs_reg_t **regs;
 }regs_ch_t;
 
 typedef struct LIBLTE_API {
   lte_cell_t cell;
-  uint8_t max_ctrl_symbols;
-  uint8_t cfi;
+  uint32_t max_ctrl_symbols;
+  uint32_t cfi;
   bool cfi_initiated;
-  uint8_t ngroups_phich;
+  uint32_t ngroups_phich;
   
   phich_resources_t phich_res;
   phich_length_t phich_len;
@@ -71,7 +71,7 @@ typedef struct LIBLTE_API {
   regs_ch_t pdcch[3]; /* PDCCH indexing, permutation and interleaving is computed for
             the three possible CFI value */
   
-  uint16_t nof_regs;
+  uint32_t nof_regs;
   regs_reg_t *regs;
 }regs_t;
 
@@ -82,10 +82,10 @@ LIBLTE_API int regs_init(regs_t *h,
 
 LIBLTE_API void regs_free(regs_t *h);
 LIBLTE_API int regs_set_cfi(regs_t *h, 
-                            uint8_t nof_ctrl_symbols);
+                            uint32_t nof_ctrl_symbols);
 
 
-LIBLTE_API uint16_t regs_pcfich_nregs(regs_t *h);
+LIBLTE_API uint32_t regs_pcfich_nregs(regs_t *h);
 LIBLTE_API int regs_pcfich_put(regs_t *h, 
                                cf_t pcfich_symbols[REGS_PCFICH_NSYM], 
                                cf_t *slot_symbols);
@@ -94,22 +94,22 @@ LIBLTE_API int regs_pcfich_get(regs_t *h,
                                cf_t *slot_symbols, 
                                cf_t pcfich_symbols[REGS_PCFICH_NSYM]);
 
-LIBLTE_API uint16_t regs_phich_nregs(regs_t *h);
+LIBLTE_API uint32_t regs_phich_nregs(regs_t *h);
 LIBLTE_API int regs_phich_add(regs_t *h, 
                               cf_t phich_symbols[REGS_PHICH_NSYM], 
-                              uint8_t ngroup, 
+                              uint32_t ngroup, 
                               cf_t *slot_symbols);
 
 LIBLTE_API int regs_phich_get(regs_t *h, 
                               cf_t *slot_symbols, 
                               cf_t phich_symbols[REGS_PHICH_NSYM], 
-                              uint8_t ngroup);
+                              uint32_t ngroup);
 
-LIBLTE_API uint8_t regs_phich_ngroups(regs_t *h);
+LIBLTE_API uint32_t regs_phich_ngroups(regs_t *h);
 LIBLTE_API int regs_phich_reset(regs_t *h, 
                                 cf_t *slot_symbols);
 
-LIBLTE_API int regs_pdcch_nregs(regs_t *h, uint8_t cfi);
+LIBLTE_API int regs_pdcch_nregs(regs_t *h, uint32_t cfi);
 LIBLTE_API int regs_pdcch_put(regs_t *h, 
                               cf_t *pdcch_symbols, 
                               cf_t *slot_symbols);
